requirement.txt
absl-py==0.11.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dlib==19.21.1
entrypoints==0.3
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.32.0
h5py==2.10.0
idna==2.6
imageio==2.9.0
imgaug==0.4.0
imutils==0.5.4
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mtcnn==0.1.0
mxnet==1.7.0.post2
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.2.0
numpy==1.16.6
oauthlib==3.1.0
opencv-python==4.5.1.48
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pathlib==1.0.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclust==0.2.0
pyclustering==0.10.1.2
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
pyqt5-tools==5.15.2.3.0.2
pyrsistent==0.17.3
PySide6==6.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.18.4
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.1
Send2Trash==1.5.0
Shapely==1.7.1
shiboken6==6.1.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
treelib==1.6.1
typing-extensions==3.7.4.3
urllib3==1.22
utils==1.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.4.3
